Refactors Drawer into common components
Refactors browser logic into separate files

Moves Paginator into common components

Makes SearchInput a generic component
