+++ src/addons/runs/History.svelte
@@ -7,7 +7,7 @@
   import HistoryEvent, { type Run } from "./HistoryEvent.svelte";
   import History24 from "svelte-octicons/lib/History24.svelte";
   import Loader from "../../common/Loader.svelte";
-  import Paginator from "../Paginator.svelte";
+  import Paginator from "../../common/Paginator.svelte";
   import { baseApiUrl } from "../../api/base.js";
 
   let res: {
