+++ src/addons/runs/Scheduled.svelte
@@ -6,7 +6,7 @@
   import Clock24 from "svelte-octicons/lib/Clock24.svelte";
   import Error from "../../common/icons/Error.svelte";
   import Loader from "../../common/Loader.svelte";
-  import Paginator from "../Paginator.svelte";
+  import Paginator from "../../common/Paginator.svelte";
   import ScheduledEvent, { type Event } from "./ScheduledEvent.svelte";
 
   import { baseApiUrl } from "../../api/base.js";
